Artillery some fixses

+++ Assets/Sourse/AI/EnemyAITemplate.cs
@@ -9,7 +9,7 @@ public class EnemyAITemplate : ScriptableObject
     [SerializeField] private float _moveDelay;
     [SerializeField] private float _chanceBuyExpensiveSoldier;
     [SerializeField] private float _chanceBuyRandomSoldier;
-    [SerializeField] private float _chanceShootArtillery;
+    [SerializeField] private int _chanceShootArtillery;
 
     public float MoveDelay => _moveDelay;
     public float TickDelay => _tickDelay;
@@ -17,5 +17,5 @@ public class EnemyAITemplate : ScriptableObject
     public float FinalAttackImpact => _finalAttackImpact;
     public float ChanceBuyExpensiveSoldier => _chanceBuyExpensiveSoldier;
     public float ChanceBuyRandomSoldier => _chanceBuyRandomSoldier;
-    public float ChanceShootArtillery => _chanceShootArtillery;
+    public int ChanceShootArtillery => _chanceShootArtillery;
 }
